gcc: silence sign compare warnings
This commit silences the following gcc sign compare errors.
dvb/dvb.cpp:1735:11: warning: comparison of integer expressions of different signedness: ‘off_t’ {aka ‘long int’} and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
dvb/decoder.cpp:1271:15: warning: comparison of integer expressions of different signedness: ‘size_t’ {aka ‘long unsigned int’} and ‘__off_t’ {aka ‘long int’} [-Wsign-compare]
base/rawfile.cpp:79:32: warning: comparison of integer expressions of different signedness: ‘long unsigned int’ and ‘off_t’ {aka ‘long int’} [-Wsign-compare]
base/cachedtssource.cpp:26:49: warning: comparison of integer expressions of different signedness: ‘long unsigned int’ and ‘off_t’ {aka ‘long int’} [-Wsign-compare]

+++ lib/dvb/decoder.cpp
@@ -1268,7 +1268,7 @@ RESULT eTSMPEGDecoder::showSinglePic(const char *filename)
 					eDebug("[eTSMPEGDecoder] VIDEO_CONTINUE: %m");
 				if (ioctl(m_video_clip_fd, VIDEO_CLEAR_BUFFER) < 0)
 					eDebug("[eTSMPEGDecoder] VIDEO_CLEAR_BUFFER: %m");
-				while(pos <= (s.st_size-4) && !(seq_end_avail = (!iframe[pos] && !iframe[pos+1] && iframe[pos+2] == 1 && iframe[pos+3] == 0xB7)))
+				while(pos <= static_cast<size_t>(s.st_size-4) && !(seq_end_avail = (!iframe[pos] && !iframe[pos+1] && iframe[pos+2] == 1 && iframe[pos+3] == 0xB7)))
 					++pos;
 				if ((iframe[3] >> 4) != 0xE) // no pes header
 					writeAll(m_video_clip_fd, pes_header, sizeof(pes_header));
